updated environment. tests and mypyp pass. refactor csv functs
